🚧 Loaded graph into graph-tool
