+++ urbanroute/urbanroute/geospatial/__init__.py
@@ -1,5 +1,8 @@
 """For geospatial queries and operations on graph."""
 
+from .intersection import RoadQuery
 from .intersection import update_cost
 from .ellipses import ellipse_bounding_box
 from .coord_match import coord_match
+
+__all__ = ["coord_match", "ellipse_bounding_box", "update_cost", "RoadQuery"]
